+++ packages/perseus/src/server-item-renderer.tsx
@@ -20,13 +20,8 @@ import Renderer from "./renderer";
 import Util from "./util";
 
 import type {KeypadProps} from "./mixins/provide-keypad";
-import type {
-    APIOptions,
-    KEScore,
-    FocusPath,
-    RendererInterface,
-    PerseusDependenciesV2,
-} from "./types";
+import type {APIOptions, FocusPath, PerseusDependenciesV2} from "./types";
+import type {RendererInterface, KEScore} from "@khanacademy/perseus-core";
 
 const {mapObject} = Objective;