+++ packages/perseus-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/perseus-core
 
+## 0.1.1
+
+### Patch Changes
+
+-   57f75510: Commented RendererInterface
+
+## 0.1.0
+
+### Minor Changes
+
+-   b4c06409: Add perseus-core to changeset
+
 ## 0.0.2
 
 ### Patch Changes